requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
boto3==1.22.2
botocore==1.25.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==37.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-storages==1.12.3
django-urls==1.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
msgpack==1.0.3
Pillow==9.1.0
psycopg==3.0.11
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
stripe==2.74.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
zope.interface==5.4.0
